Use `CommittedDate` as a fallback if `PushedDate` is not available.
The fix for #22 makes github-pr-resource skip PRs if the the latest
commit in a PR pushed has a later date than one in PR being pushed.
This fix works around this by using `PushedDate` by default and
resorting to `CommitDate` iff `PushedDate` is not available.

This should bring the best of both worlds -- checks for all PRs on
master and a chance for non-master-based PRs to be checked.

Signed-off-by: Milan Plzik <milan.plzik@ceai.io>